Add filter-result information for dead ECAL-cell filters.
Two new boolean variables,
- PassesECALDeadCellBEFilter
- PassesECALDeadCellBEFilter
are added to the ntuple. They are 'true' by default and the tree maker runs as before if nothing else is done. To run the filters (will not reject the event) and store the filter decision, check out the filter tool from RecoMET/METFilters:

cvs co RecoMET/METFilters

The highest current tag is lhx_13APR2012_v3. An example how to use and configure the filters is given in 'Calibration/CalibTreeMaker/python/runTreeMaker_cff.py'.